+++ src/providers/BaseViewProvider.ts
@@ -142,22 +142,6 @@ export abstract class BaseViewProvider implements vscode.WebviewViewProvider {
         vscode.commands.executeCommand("beads.openBeadDetails", message.beadId);
         break;
 
-      case "showDoltStatus":
-        vscode.commands.executeCommand("beads.showDoltStatus");
-        break;
-
-      case "startDoltServer":
-        vscode.commands.executeCommand("beads.startDoltServer");
-        break;
-
-      case "stopDoltServer":
-        vscode.commands.executeCommand("beads.stopDoltServer");
-        break;
-
-      case "openDoltLog":
-        vscode.commands.executeCommand("beads.openDoltLog");
-        break;
-
       case "openProjectFolder": {
         const project = this.projectManager.getActiveProject();
         if (project) {